Remove redundant DailyPushTestController and consolidate test APIs
- Delete DailyPushTestController.cs to avoid route conflicts
- Keep test endpoints in DailyPushController only
- Add GET /api/push/test/timezones endpoint for available timezones
- Simplify API structure with single controller approach
